+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # 1.4.13 - 23 Oct 2025
+Improvement:
+- [#1453](https://github.com/elysiajs/elysia/issues/1453) add `allowUnsafeValidationDetails` for disabling unsafe validation details in production mode
+
+Bug fix:
 - [#1502](https://github.com/elysiajs/elysia/issues/1502) afterHandle doesn't update status
 - [#1495](https://github.com/elysiajs/elysia/pull/1495) request server hook parameters are typed as any (Bun 1.3.0)
 - [#1483](https://github.com/elysiajs/elysia/pull/1483) handle standard schema validators in ValidationError.all
 - [#1459](https://github.com/elysiajs/elysia/pull/1459) fix strictPath behavior when aot: false is set
 - [#1455](https://github.com/elysiajs/elysia/pull/1455) graceful shutdown not awaiting server.stop
-- [#1499](https://github.com/elysiajs/elysia/pull/1449) fails when mergint with t.Optional schema
+- [#1499](https://github.com/elysiajs/elysia/pull/1449) fails when merging with t.Optional schema
+
+change:
+- make `@types/bun` an optional dependency
 
 # 1.4.12 - 14 Oct 2025
 Improvement: